+++ lib/Models/LeafletAttribution.ts
@@ -1,14 +1,13 @@
-import Terria from "./Terria";
-import { LeafletCredits } from "../ReactViews/Credits";
+import { IObservableArray, runInAction, observable } from "mobx";
 import L from "leaflet";
-import React from "react";
-import ReactDOM from "react-dom";
+import Terria from "./Terria";
 
 export class LeafletAttribution extends L.Control.Attribution {
-  private readonly terria: Terria;
+  private readonly _attributions: Record<string, number>;
   private map?: L.Map;
   private _container!: HTMLElement;
-  private currentAttibution?: HTMLElement;
+  @observable
+  dataAttributions: IObservableArray<string>;
 
   constructor(terria: Terria) {
     const options: L.Control.AttributionOptions = {
@@ -19,7 +18,8 @@ export class LeafletAttribution extends L.Control.Attribution {
     }
     super(options);
 
-    this.terria = terria;
+    this._attributions = {};
+    this.dataAttributions = observable([]);
   }
 
   onAdd(map: L.Map) {
@@ -39,35 +39,58 @@ export class LeafletAttribution extends L.Control.Attribution {
     return this._container;
   }
 
+  onRemove() {
+    this.map = undefined;
+  }
+
   _update() {
     if (!this.map) {
       return;
     }
-    if (this.currentAttibution) {
-      this._container.removeChild(this.currentAttibution);
-    }
 
     const attribs: string[] = [];
-    //@ts-ignore
     const attributions = this._attributions;
     for (const i in attributions) {
       if (attributions[i]) {
         attribs.push(i);
       }
     }
+  }
 
-    const domElement = document.createElement("div");
-    const element = React.createElement(LeafletCredits, {
-      hideTerriaLogo: !!this.terria.configParameters.hideTerriaLogo,
-      prefix: `${this.options.prefix}`,
-      credits: this.terria.configParameters.extraCreditLinks?.slice(),
-      dataAttributions: attribs
-    });
-    ReactDOM.render(element, domElement);
-    const child = domElement.firstElementChild as HTMLElement;
-    if (child) {
-      this.currentAttibution = child;
-      this._container?.appendChild(child);
+  addAttribution(text: string) {
+    super.addAttribution(text);
+    if (this.map) {
+      runInAction(() => {
+        this.dataAttributions.push(text);
+      });
     }
+    return this;
+  }
+
+  removeAttribution(text: string) {
+    super.removeAttribution(text);
+    if (this.map) {
+      runInAction(() => {
+        this.dataAttributions.remove(text);
+      });
+    }
+    return this;
+  }
+
+  get attributions(): string[] {
+    const attributionsList: string[] = [];
+    const attributions = this._attributions;
+    for (const i in attributions) {
+      if (attributions[i]) {
+        attributionsList.push(i);
+      }
+    }
+
+    return attributionsList;
+  }
+
+  get prefix(): string | undefined {
+    if (!this.options.prefix) return undefined;
+    return `${this.options.prefix}`;
   }
 }