+++ lib/ReactViews/Credits/Cesium/CesiumCredits.props.ts
@@ -1,8 +1,0 @@
-import { ICredit } from "../Credit.type";
-
-export interface ICesiumCreditsProps {
-  hideTerriaLogo: boolean;
-  credits?: ICredit[];
-  expandDataCredits?: () => void;
-  cesiumLogoElement?: HTMLElement;
-}