+++ lib/ReactViews/Credits/Leaflet/LeafletCredits.props.ts
@@ -1,8 +1,0 @@
-import { ICredit } from "../Credit.type";
-
-export interface ILeafletCreditsProps {
-  hideTerriaLogo: boolean;
-  prefix?: string;
-  credits?: ICredit[];
-  dataAttributions: string[];
-}